Edit profile page
This is an example of a more detailed commit message which should come after the first subject line. Explain the problem that this commit is solving. Focus on why you are making this change as opposed to how (the code explains that). Are there side effects or other unintuitive consequences of this change? Here's the place to explain them.

Further paragraphs come after blank lines.

- Bullet points can be preceded by hyphens

- Empty line should come between

Issue trackers should and reference should be put at the pottom:

Resovles: #123
See also: #3213, #223